Exploring linked list
A following question is answered using linked list data structure

Create an Employee record that consists of Employee ID, Employee name, and Employee's age. The data for the creation of the record is given below: {{Employee ID: 1001, Employee Name: B, age:27}, {ID:1003, Name: D, age: 20}, {ID:1004, Name: E, age: 25}}. Write a menu-driven program to perform the following operations

Insert a new record at the beginning of the list.
Insert a new record at the end of the list.
Insert a new record next to the record with a specific Employee ID.
Delete the Employee Record for a specific Employee ID.
Sort the record based on the Employee age and display it.
Delete the last record in the current sorted list.